backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 文件解析
pdfplumber==0.10.3
python-docx==1.1.0
PyPDF2==3.0.1
extract-msg==0.45.0
python-magic-bin==0.4.14  # Windows下文件类型识别

# AI模型调用（智谱AI）
zhipuai==2.0.1

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 工具库
pydantic==2.5.3
python-dotenv==1.0.0
